moorcheh-service/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
moorcheh-sdk>=0.1.0
python-dotenv==1.0.1
pydantic==2.10.4
